requirements_dev.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
astroid==3.2.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
coverage==7.5.1
cycler==0.12.1
dill==0.3.8
docutils==0.21.2
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
furo==2024.5.6
grpcio==1.63.0
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
inplace-abn==1.1.0
isort==5.13.2
Jinja2==3.1.4
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lightning==2.2.4
lightning-utilities==0.11.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pretty-errors==1.2.25
protobuf==5.26.1
psutil==5.9.8
Pygments==2.18.0
pylint==3.2.1
pyparsing==3.1.2
pytest==8.2.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tomlkit==0.12.5
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
